:truck: Rename: 소셜 인증 관련 Result 이동
- 소셜 인증 관련 Result 객체를 response에서 record 패키지로 이동
