+++ src/main/java/kr/modusplant/domains/account/social/adapter/SocialIdentityTokenHelper.java
@@ -3,7 +3,7 @@ package kr.modusplant.domains.account.social.adapter;
 import io.jsonwebtoken.Claims;
 import kr.modusplant.domains.account.social.domain.vo.enums.SocialProvider;
 import kr.modusplant.domains.account.social.usecase.record.TempTokenInfo;
-import kr.modusplant.domains.account.social.usecase.response.SocialPendingResult;
+import kr.modusplant.domains.account.social.usecase.record.SocialPendingResult;
 import kr.modusplant.infrastructure.jwt.provider.JwtTokenProvider;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
