+++ src/main/java/kr/modusplant/domains/account/social/adapter/controller/SocialIdentityController.java
@@ -14,10 +14,10 @@ import kr.modusplant.domains.account.social.usecase.port.mapper.SocialIdentityMa
 import kr.modusplant.domains.account.social.usecase.port.repository.SocialIdentityRepository;
 import kr.modusplant.domains.account.social.usecase.record.TempTokenInfo;
 import kr.modusplant.domains.account.social.usecase.request.SocialSignUpRequest;
-import kr.modusplant.domains.account.social.usecase.response.LoginResult;
-import kr.modusplant.domains.account.social.usecase.response.NeedLinkResult;
-import kr.modusplant.domains.account.social.usecase.response.NeedSignupResult;
-import kr.modusplant.domains.account.social.usecase.response.SocialLoginResult;
+import kr.modusplant.domains.account.social.usecase.record.LoginResult;
+import kr.modusplant.domains.account.social.usecase.record.NeedLinkResult;
+import kr.modusplant.domains.account.social.usecase.record.NeedSignupResult;
+import kr.modusplant.domains.account.social.usecase.record.SocialLoginResult;
 import kr.modusplant.shared.enums.AuthProvider;
 import kr.modusplant.shared.enums.Role;
 import kr.modusplant.shared.kernel.Email;
