+++ src/main/java/kr/modusplant/domains/account/social/usecase/port/mapper/SocialIdentityMapper.java
@@ -4,7 +4,7 @@ import kr.modusplant.domains.account.social.domain.vo.SocialMemberProfile;
 import kr.modusplant.domains.account.social.domain.vo.SocialProfile;
 import kr.modusplant.domains.account.social.domain.vo.enums.SocialProvider;
 import kr.modusplant.domains.account.social.usecase.port.client.dto.SocialUserInfo;
-import kr.modusplant.domains.account.social.usecase.response.LoginResult;
+import kr.modusplant.domains.account.social.usecase.record.LoginResult;
 import kr.modusplant.shared.enums.AuthProvider;
 
 public interface SocialIdentityMapper {
