+++ src/main/java/kr/modusplant/domains/account/social/usecase/record/LoginResult.java
@@ -1,4 +1,4 @@
-package kr.modusplant.domains.account.social.usecase.response;
+package kr.modusplant.domains.account.social.usecase.record;
 
 import kr.modusplant.shared.enums.Role;
