+++ src/main/java/kr/modusplant/domains/account/social/usecase/record/NeedLinkResult.java
@@ -1,4 +1,4 @@
-package kr.modusplant.domains.account.social.usecase.response;
+package kr.modusplant.domains.account.social.usecase.record;
 
 import kr.modusplant.domains.account.social.domain.vo.enums.SocialProvider;
