+++ src/main/java/kr/modusplant/domains/account/social/usecase/record/SocialLoginResult.java
@@ -1,4 +1,4 @@
-package kr.modusplant.domains.account.social.usecase.response;
+package kr.modusplant.domains.account.social.usecase.record;
 
 public sealed interface SocialLoginResult permits LoginResult, SocialPendingResult {
 }
