+++ src/test/java/kr/modusplant/domains/account/social/framework/in/web/rest/SocialIdentityRestControllerTest.java
@@ -11,9 +11,9 @@ import kr.modusplant.domains.account.social.usecase.enums.OAuthType;
 import kr.modusplant.domains.account.social.usecase.record.TempTokenInfo;
 import kr.modusplant.domains.account.social.usecase.request.SocialAuthRequest;
 import kr.modusplant.domains.account.social.usecase.request.SocialSignUpRequest;
-import kr.modusplant.domains.account.social.usecase.response.LoginResult;
-import kr.modusplant.domains.account.social.usecase.response.NeedLinkResult;
-import kr.modusplant.domains.account.social.usecase.response.NeedSignupResult;
+import kr.modusplant.domains.account.social.usecase.record.LoginResult;
+import kr.modusplant.domains.account.social.usecase.record.NeedLinkResult;
+import kr.modusplant.domains.account.social.usecase.record.NeedSignupResult;
 import kr.modusplant.domains.account.social.usecase.response.SocialLoginResponse;
 import kr.modusplant.framework.jackson.holder.ObjectMapperHolder;
 import kr.modusplant.framework.jackson.http.response.DataResponse;
